Create closeWelcomeModal, goToPageEditor and goToPostEditor utils in PlayWright e2e tests
