+++ plugins/woocommerce/tests/e2e-pw/tests/merchant/command-palette.spec.js
@@ -1,35 +1,10 @@
 const { test, expect } = require( '@playwright/test' );
 const wcApi = require( '@woocommerce/woocommerce-rest-api' ).default;
+const { goToPostEditor } = require( '../../utils/editor' );
 
 // need to figure out whether tests are being run on a mac
 const macOS = process.platform === 'darwin';
 const cmdKeyCombo = macOS ? 'Meta+k' : 'Control+k';
-
-const goToPostEditor = async ( { page } ) => {
-	await page.goto( 'wp-admin/post-new.php' );
-
-	const welcomeModalVisible =
-		await test.step( 'Check if the Welcome modal appeared', async () => {
-			return await page
-				.getByRole( 'heading', {
-					name: 'Welcome to the block editor',
-				} )
-				.isVisible();
-		} );
-
-	if ( welcomeModalVisible ) {
-		await test.step( 'Welcome modal appeared. Close it.', async () => {
-			await page
-				.getByRole( 'document' )
-				.getByRole( 'button', { name: 'Close' } )
-				.click();
-		} );
-	} else {
-		await test.step( 'Welcome modal did not appear.', async () => {
-			// do nothing.
-		} );
-	}
-};
 
 const clickOnCommandPaletteOption = async ( { page, optionName } ) => {
 	// Press `Ctrl` + `K` to open the command palette.
