+++ plugins/woocommerce/tests/e2e-pw/tests/shopper/product-tags-attributes.spec.js
@@ -1,6 +1,7 @@
 const { test, expect, request } = require( '@playwright/test' );
 const { admin } = require( '../../test-data/data' );
 const pageTitle = 'Product Showcase';
+const { goToPageEditor } = require( '../../utils/editor' );
 const wcApi = require( '@woocommerce/woocommerce-rest-api' ).default;
 
 const singleProductPrice1 = '5.00';
@@ -265,17 +266,7 @@ test.describe( 'Browse product tags and attributes from the product page', () =>
 
 	test( 'can see products showcase', async ( { page } ) => {
 		// create as a merchant a new page with Products block
-		await page.goto( 'wp-admin/post-new.php?post_type=page' );
-
-		const welcomeModalVisible = await page
-			.getByRole( 'heading', {
-				name: 'Welcome to the block editor',
-			} )
-			.isVisible();
-
-		if ( welcomeModalVisible ) {
-			await page.getByRole( 'button', { name: 'Close' } ).click();
-		}
+		await goToPageEditor( { page } );
 
 		await page
 			.getByRole( 'textbox', { name: 'Add Title' } )
